+++ db/schema/09_membrships.sql
@@ -7,5 +7,5 @@ CREATE TABLE memberships (
   available BOOLEAN NOT NULL DEFAULT TRUE,
   promotion_code VARCHAR(255),
   is_unlimited BOOLEAN NOT NULL DEFAULT FALSE,
-  session_num TINYINT
+  session_num SMALLINT
 );
